pyproject.toml
[project]
name = "jobs-agent"
version = "0.1.0"
description = "Job aggregation and analysis agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.4",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "httpx[socks]>=0.28.1",
    "openai>=1.0.0",
    "boto3>=1.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
